Refactor the serial signing API
